// src/app/core/models/checkout.ts
export class Checkout {
    id: number;
    firstname: string;
    lastname: string;
    mobile: number;
    email: string;
    address1: string;
    address2: string;
    city: string;
    postalcode: number;
    country: string;
    state: string;
    updated: Date;
}

export class PaymentMethod{
    name: string;
    value: string;
}
    

    